Fixed issue where changing the story's priority would switch to the Backlog details pane that the story belonged to. Updated Story list extractor (to monitor for changes in priority also). Implemented priority changing when re-ordered within the ScrumBoard. General bug fixes relating to the ScrumBoard. story:284

+++ src/main/java/com/thirstygoat/kiqo/gui/sprint/StoryRowViewModel.java
@@ -8,10 +8,7 @@ import com.thirstygoat.kiqo.model.Task;
 import com.thirstygoat.kiqo.util.GoatModelWrapper;
 import de.saxsys.mvvmfx.ViewModel;
 import javafx.beans.binding.Bindings;
-import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.SimpleIntegerProperty;
-import javafx.beans.property.SimpleStringProperty;
-import javafx.beans.property.StringProperty;
+import javafx.beans.property.*;
 import javafx.collections.FXCollections;
 import javafx.collections.ListChangeListener;
 import javafx.collections.ObservableList;
@@ -114,7 +111,7 @@ public class StoryRowViewModel implements Loadable<Story>, ViewModel {
         return description;
     }
 
-    public int getPriority() {
+    public float getPriority() {
         return priority.get();
     }
